user type completed

+++ database/migrations/2024_07_09_141516_add_type_to_letters_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::table('letters', function (Blueprint $table) {
             $table->after('short_title', function (Blueprint $table) {
-                $table->tinyInteger('type')->comment('1: official, 2: civil');
+                $table->integer('type')->comment('1: official, 2: civil');
             });
         });
     }
